fixup! Read PG vars in create_repository.sh
